refactor(storage): split ts-storage into abstraction + ts-storage-duckdb impl
ts-storage now contains only the StorageBackend trait, query DTOs, and the
generic buffer/sink/retention scaffolding. The DuckDB implementation moves
to a new ts-storage-duckdb crate. The multi-backend `create_backend` dispatch
function moves to app/tokenscope (the assembly layer), so backend impl
crates remain ignorant of each other.

Why: ts-api production code only uses the trait + DTOs, but transitively
linked libduckdb (~64 MB) + arrow (10 sub-crates) because ts-storage
bundled both abstraction and impl. CLAUDE.md already declares "pluggable
backend via trait" with planned PostgreSQL / ClickHouse backends — this
is the standard sqlx-core / sqlx-postgres / sqlx-mysql shape.

Impact:
- `cargo tree -p ts-api -e normal` no longer references duckdb/arrow.
- ts-api/tests/pcap_extract_route.rs integration test stops linking
  libduckdb (5 → 4 duckdb-linked test bins workspace-wide).
- ts-storage changes no longer trigger ts-api re-link unless the trait
  or DTO surface changes.
- Adding ts-storage-postgres later is a one-arm extension to
  tokenscope::create_backend, no fan-in to backend crates.

No behavior change; 301 workspace tests pass.
