+++ server/Cargo.toml
@@ -62,6 +62,7 @@ ts-llm = { path = "ts-llm" }
 ts-turn = { path = "ts-turn" }
 ts-metrics = { path = "ts-metrics" }
 ts-storage = { path = "ts-storage" }
+ts-storage-duckdb = { path = "ts-storage-duckdb" }
 ts-pcap-extract = { path = "ts-pcap-extract" }
 axum = "0.8"
 tower-http = { version = "0.6", features = ["cors"] }
